изменение кнопки инфо
