+++ Navigation/SceneDelegate.swift
@@ -1,9 +1,7 @@
 //
 //  SceneDelegate.swift
 //  Navigation
-//
-//  Created by Мамуля on 03.06.2023.
-//
+
 
 import UIKit
