+++ Profile/ProfileViewController.swift
@@ -2,18 +2,28 @@
 //  ProfileViewController.swift
 //  Navigation
 //
-//  Created by Мамуля on 10.06.2023.
 //
 
 import UIKit
 
 class ProfileViewController: UIViewController {
     
+    let profileHeaderView = ProfileHeaderView()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
         view.backgroundColor = .lightGray
+        view.addSubview(profileHeaderView)
+        
+        
     }
+    
+    override func viewWillLayoutSubviews() {
+        super.viewWillLayoutSubviews()
+        profileHeaderView.frame = view.frame
+    }
+    
     
 }
